[OSIS-3776] move charge repartition views to attribution app
	new file:   attribution/tests/views/charge_repartition/__init__.py
	new file:   attribution/tests/views/charge_repartition/common.py
	renamed:    base/tests/views/learning_units/test_charge_repartition.py -> attribution/tests/views/charge_repartition/test_create.py
	new file:   attribution/tests/views/charge_repartition/test_update.py
	new file:   attribution/tests/views/learning_unit/__init__.py
	new file:   attribution/tests/views/learning_unit/test_create.py
	new file:   attribution/tests/views/learning_unit/test_delete.py
	new file:   attribution/tests/views/learning_unit/test_update.py
	modified:   attribution/tests/views/test_attribution.py
	modified:   attribution/urls.py
	renamed:    base/views/learning_units/attribution.py -> attribution/views/attribution.py
	renamed:    base/views/learning_units/charge_repartition.py -> attribution/views/charge_repartition/create.py
	new file:   attribution/views/charge_repartition/update.py
	modified:   attribution/views/learning_unit/common.py
	modified:   attribution/views/learning_unit/create.py
	modified:   attribution/views/learning_unit/delete.py
	modified:   attribution/views/learning_unit/update.py
	modified:   base/forms/learning_unit/attribution_charge_repartition.py
	modified:   base/tests/views/learning_units/test_attribution.py
	modified:   base/urls.py
