+++ attribution/tests/views/charge_repartition/test_create.py
@@ -33,71 +33,12 @@ from attribution.models.attribution_charge_new import AttributionChargeNew
 from attribution.models.attribution_new import AttributionNew
 from attribution.tests.factories.attribution_charge_new import AttributionChargeNewFactory
 from attribution.tests.factories.attribution_new import AttributionNewFactory
+from attribution.tests.views.charge_repartition.common import TestChargeRepartitionMixin
 from base.tests.factories.learning_component_year import LecturingLearningComponentYearFactory, \
     PracticalLearningComponentYearFactory
 from base.tests.factories.learning_unit_year import LearningUnitYearFullFactory, LearningUnitYearPartimFactory
 from base.tests.factories.person import PersonWithPermissionsFactory
 from base.views.mixins import RulesRequiredMixin
-
-
-class TestChargeRepartitionMixin:
-    @classmethod
-    def setUpTestData(cls):
-        cls.learning_unit_year = LearningUnitYearPartimFactory()
-        cls.lecturing_component = LecturingLearningComponentYearFactory(learning_unit_year=cls.learning_unit_year)
-        cls.practical_component = PracticalLearningComponentYearFactory(learning_unit_year=cls.learning_unit_year)
-
-        cls.full_learning_unit_year = LearningUnitYearFullFactory(
-            learning_container_year=cls.learning_unit_year.learning_container_year,
-            academic_year=cls.learning_unit_year.academic_year
-        )
-        cls.lecturing_component_full = LecturingLearningComponentYearFactory(
-            learning_unit_year=cls.full_learning_unit_year
-        )
-        cls.practical_component_full = PracticalLearningComponentYearFactory(
-            learning_unit_year=cls.full_learning_unit_year
-        )
-        cls.person = PersonWithPermissionsFactory('can_access_learningunit')
-
-    def setUp(self):
-        self.attribution = AttributionNewFactory(
-            learning_container_year=self.learning_unit_year.learning_container_year
-        )
-        attribution_id = self.attribution.id
-        self.charge_lecturing = AttributionChargeNewFactory(
-            attribution=self.attribution,
-            learning_component_year=self.lecturing_component
-        )
-        self.charge_practical = AttributionChargeNewFactory(
-            attribution=self.attribution,
-            learning_component_year=self.practical_component
-        )
-
-        self.attribution_full = self.attribution
-        self.attribution_full.id = None
-        self.attribution_full.save()
-        self.charge_lecturing_full = AttributionChargeNewFactory(
-            attribution=self.attribution_full,
-            learning_component_year=self.lecturing_component_full
-        )
-        self.charge_practical_full = AttributionChargeNewFactory(
-            attribution=self.attribution_full,
-            learning_component_year=self.practical_component_full
-        )
-
-        self.attribution = AttributionNew.objects.get(id=attribution_id)
-        self.client.force_login(self.person.user)
-
-        self.patcher = patch.object(RulesRequiredMixin, "test_func", return_value=True)
-        self.mocked_permission_function = self.patcher.start()
-
-    def tearDown(self):
-        self.patcher.stop()
-
-    def clean_partim_charges(self):
-        self.charge_practical.delete()
-        self.charge_lecturing.delete()
-        self.attribution.delete()
 
 
 class TestSelectAttributionView(TestChargeRepartitionMixin, TestCase):
@@ -185,72 +126,5 @@ class TestAddChargeRepartition(TestChargeRepartitionMixin, TestCase):
         attribution_partim = AttributionNew.objects.exclude(id=self.attribution_full.id).get(tutor=self.attribution_full.tutor)
         self.assertNotEqual(attribution_partim.external_id, self.attribution_full.external_id)
         self.assertIsNone(attribution_partim.external_id)
-        self.assertRedirects(response,
-                             reverse("learning_unit_attributions", args=[self.learning_unit_year.id]))
-
-
-class TestEditChargeRepartition(TestChargeRepartitionMixin, TestCase):
-    def setUp(self):
-        super().setUp()
-        self.url = reverse("edit_charge_repartition", args=[self.learning_unit_year.id, self.attribution.id])
-
-    def test_login_required(self):
-        self.client.logout()
-
-        response = self.client.get(self.url)
-        self.assertRedirects(response,  '/login/?next={}'.format(self.url))
-
-    def test_template_used_with_get(self):
-        response = self.client.get(self.url)
-
-        self.assertTrue(self.mocked_permission_function.called)
-        self.assertEqual(response.status_code, HttpResponse.status_code)
-        self.assertTemplateUsed(response, "learning_unit/add_charge_repartition_inner.html")
-
-    def test_post(self):
-        data = {
-            'lecturing_form-allocation_charge': 50,
-            'practical_form-allocation_charge': 10
-        }
-        response = self.client.post(self.url, data=data)
-
-        self.charge_lecturing.refresh_from_db()
-        self.charge_practical.refresh_from_db()
-        self.assertEqual(self.charge_lecturing.allocation_charge, 50)
-        self.assertEqual(self.charge_practical.allocation_charge, 10)
-
-        self.assertRedirects(response,
-                             reverse("learning_unit_attributions", args=[self.learning_unit_year.id]))
-
-
-class TestRemoveChargeRepartition(TestChargeRepartitionMixin, TestCase):
-    def setUp(self):
-        super().setUp()
-        self.url = reverse("remove_attribution", args=[self.learning_unit_year.id, self.attribution.id])
-
-    def test_login_required(self):
-        self.client.logout()
-
-        response = self.client.get(self.url)
-        self.assertRedirects(response,  '/login/?next={}'.format(self.url))
-
-    def test_template_used_with_get(self):
-        response = self.client.get(self.url)
-
-        self.assertTrue(self.mocked_permission_function.called)
-        self.assertEqual(response.status_code, HttpResponse.status_code)
-        self.assertTemplateUsed(response, "learning_unit/remove_charge_repartition_confirmation_inner.html")
-
-    def test_delete_data(self):
-        response = self.client.delete(self.url)
-
-        self.assertFalse(AttributionNew.objects.filter(id=self.attribution.id).exists())
-        self.assertFalse(
-            AttributionChargeNew.objects.filter(id__in=(self.charge_lecturing.id, self.charge_practical.id)).exists()
-        )
-
-    def test_delete_redirection(self):
-        response = self.client.delete(self.url, follow=False)
-
         self.assertRedirects(response,
                              reverse("learning_unit_attributions", args=[self.learning_unit_year.id]))
